Flask_server/requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
nose==1.3.7
psycopg2-binary==2.8.4
Werkzeug==0.16.1
pyjwt==2.3.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
pandas==1.4.2
pandas==1.4.2
numpy==1.22.2
python-dateutil==2.8.2
pytz==2021.3
scipy==1.8.0
six==1.16.0